Foundation: make (cross-)compile without errors but still thousands of warnings, bugs and disabled code sections because they need to become compatible to the new ObjC-API (libobjc.so.4)

+++ Foundation/Sources/NSString.h
@@ -170,7 +170,7 @@ extern NSString *NSParseErrorException;
 - (NSArray *) componentsSeparatedByCharactersInSet:(NSCharacterSet *) set;
 - (NSArray *) componentsSeparatedByString:(NSString *) separator;
 - (const char *) cString;								// C Strings
-- (unsigned int) cStringLength;
+- (NSUInteger) cStringLength;
 - (const char *) cStringUsingEncoding:(NSStringEncoding) encoding;
 - (NSData *) dataUsingEncoding:(NSStringEncoding) encoding;
 - (NSData *) dataUsingEncoding:(NSStringEncoding) encoding
